+++ core/src/main/java/com/shatteredpixel/yasd/general/items/wands/WandOfDamnation.java
@@ -74,7 +74,7 @@ public class WandOfDamnation extends Wand {
         Char ch = defender;
         DeferredDeath buff = ch.buff(DeferredDeath.class);
         if (buff != null) {
-            buff.recover(Random.Int(staff.power()*2+2));
+            buff.recover(Random.Int((int) staff.power()*2+2));
         }
     }
